RM-2051 UI integration test for classify document.
Currently just created a test for classify document (classify record to
come next. Added this test in a new test set in the testng.xml file. Group
the data setup methods so that they also get run.

Added waitForInvisibility helper method that takes a WebElement.

Created a ClassifyContentDialog page object used by the test. Note that
this contains a mixture of css and xpath. I've tried to prefer css
selectors, but in places found I needed to match the text node and
couldn't get this to work using css.

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@102728 7033f55c-6c29-4ad8-afba-765cb616e738