Re-enable classification UI integration test.
Create separate collaboration site for use by the sanity and integration
tests.  This was causing a failure before.

Changed method for adding a classification reason to some content slightly,
so it's now quite a bit less hacky.  (Nb. It still feels a bit hacky).

Fixed a copy/paste error in the test name for CreateCollabSite.

+review RM-39

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@102945 7033f55c-6c29-4ad8-afba-765cb616e738