RM-2051 Add remaining tests for acceptance criteria.
Support reseting state after classifying a document (currently this works
by deleting and re-uploading the document).

Don't wait for the "Create" button to be invisible on the classify dialog
as the "click" action already contains a wait in it.

Play around with the classification levels in the test data, so that we're
not classifying using the value that initially appears in the dialog.

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@103346 7033f55c-6c29-4ad8-afba-765cb616e738