RM-2272 Create automated UI test for a uncleared user.
This included making some fixes to the logout and login handling. I have
tried to ensure that at some point in the future we will be able to
parallelise the testing across several threads (which is a feature of
TestNG), but I have no idea whether this is likely to be possible in the
near term with our Bamboo set up.

Make sure the RM site is created before attempting to create RM users.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@105155 7033f55c-6c29-4ad8-afba-765cb616e738