RM-2066 Finish UI test now we can classify a document.
Also start refactoring groups so that the TestData class contains a
reference of everything that happens to each data object, and to make
it easier to find the right TestNG group to depend on.

Remove data setup method that creates a classified document, as we already
have a test that does the same thing. Use this test as dependency for
test checking classified documents can't be shared.

Generally try to tidy up test dependencies, and remove dependencies that
aren't needed.

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@105510 7033f55c-6c29-4ad8-afba-765cb616e738