RM-2272 Tidying all automated UI tests.
Finish the implementation of the unfinished tests from previous sprints.
Tidy the dependency naming, so that all dependencies are called GROUP_blah
and are defined in a sensible place in TestData.

Ensure that all tests are in the group "integration", and delete all other
groups (except those needed for dependencies). Tidy the TestNG.xml file to
just kick off the "integration" group.

Create a helper method that can navigate directly to a folder in the file
plan (as previously we clicked on every intermediate category and folder,
which takes a lot of time to render).

Remove the sanity tests from the TestNG.xml file following analysis of
what extra testing they're currently providing.  The only things that were
being tested by the sanity tests and not the integration tests were:
 - Deleting a collaboration site (this isn't really an RM feature);
 - Creating an unfiled record folder (we are testing the creation of other
   folders though).

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@105564 7033f55c-6c29-4ad8-afba-765cb616e738