RM-2132 RM-2133 Tests for different users viewing classified documents.
Add two UI automation tests that are for a 'secret' user and 'no clearance'
user viewing documents classified at different levels.

The tests don't pass at the moment as the feature isn't complete. On top of
this the tests will cause issues on the Bamboo server as they involve
changing user. Consequently the tests are ignored at the moment.

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@105780 7033f55c-6c29-4ad8-afba-765cb616e738