Change location of records to reduce test clicks.
When running a single method we should try to reduce the time taken to
set up the data.  This change moves the record folder where records are
created to be in "category one" rather than a sub-category.  The sub-
category is still created as part of the data set up, but at the moment
nothing uses it, and it's not a dependency of any test method.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@106915 7033f55c-6c29-4ad8-afba-765cb616e738