Belt and braces for logging out.
Retry until we see "Login" contained in the title. We have been seeing an
intermittent issue where Selenium attempts to logout, but somehow
completely fails (maybe because the menus aren't loaded when it tries?)
Since the screenshots always seem to show a normal page, it seems worth
retrying a couple of times, in case we can successfully log out then.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@107184 7033f55c-6c29-4ad8-afba-765cb616e738