Try waiting for FADEIN to finish when logging in.
I have no idea if this will solve the intermittent issues on Bamboo, but
we sometimes see screenshots of the login page in the middle of the fade-in.
Hopefully by waiting for the same length of time as the fade-in, and then
re-rendering, we can improve the chance of logging in successfully.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@107394 7033f55c-6c29-4ad8-afba-765cb616e738