Try to find out if webDriver is blocking by logging time.
The webDrive.get javadoc says that it will "block until the load is
complete", but in the test run #552 it looks like the test
checkCanClassifyCompleteRecord didn't manage to ever load the page, even
after issuing the webDriver.get(url) command 20 times.  Add timing
information to see if webDriver.get is blocking for any sensible length of
time.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@107539 7033f55c-6c29-4ad8-afba-765cb616e738