Try waiting for an increasing amount of time between page load attempts.
From the reporter output of run #553 it looks like loading the document
library failed 20 times in the space of a second (for the test
cantClassifyHeldRecord), and then loading the same page worked fine in the
next test (classifiedRecordIndicator) a few seconds later. Hopefully by
increasing the interval between attempts we can out-wait any issues when
the client is busy/unresponsive.

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@107570 7033f55c-6c29-4ad8-afba-765cb616e738