RM-2466 Pass through total results count as if nothing is filtered.
Change PagingResultsPostMethodInvocationProcessor to always return the
total results count as if nothing is filtered. This should fix the case
where nothing is filtered, and is a concession to the fact that we cannot
know how many total results there will be from a single page. We will fix
this later by filtering the results before we get to this processor, but
the processor is a good backup filter.

Change the rm-automation tests not to use the pagination details at the
bottom of the browse list pages, as these will often now be incorrect.

In making this change I also looked at using the getTotalResultCount()
interface with a range as the interface supports returning a minimum and
maximum on the total number of results. This would be very useful here,
as it should allow us to say that we don't know how many results there
will be.  However most of the webscripts don't support a range, and so
this leads to some odd results in Share (e.g. 'Showing results 1 to 50
of -1').

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@108482 7033f55c-6c29-4ad8-afba-765cb616e738