RM-2441, RM-2442, RM-2443 Fix timezone issue.
Previously these tests would fail when the day in the local timezone wasn't
the same as the day in UTC. Rather than check the day is correct (which
introduces a small race condition anyway), just validate that a valid date
has been set (i.e. it is not still set to '(None)').

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@109600 7033f55c-6c29-4ad8-afba-765cb616e738