RM-2227 Remove commons-logging dependency from Share amp.
This should mean we no longer need to install the AMP with "-force".

Locally the issue was fixed by changing the ${mvn.exec} definition in
build.local.properties to say:
mvn.exec=mvn
And ensuring my system version of maven is a recent version (3.2.5 rather
than 3.0.5 which I was using by accident before).

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@111551 7033f55c-6c29-4ad8-afba-765cb616e738