RM-2606 Classify the head of the version history.
As the version store does not allow editing properties or aspects, we
instead delete the head and recreate it using the classified version of the
file.

Note that classifying some content does not classify all versions, only the
current one. This currently causes other aspects to be added to the history
node, I'm not sure this is the right behaviour, but it felt like an edge
case...

+review RM

git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/modules/recordsmanagement/HEAD@113116 7033f55c-6c29-4ad8-afba-765cb616e738